requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
mcp-use==0.0.5
langchain-ollama==0.3.4
langchain-groq==0.2.5 